// src/types/Orchestration.ts
export type Orchestration = {
  id: string;
  playId: string;
  orchestrationName: string;
  orchestrationSize?: string;
  orchestrationDescription?: string;
  orchestrationType?: OrchestrationTypes;
  musicStyle?: OrchestrationMusicStyle;
  musicMoodStyle?: OrchestrationMusicMoodStyle;
  musicGenrePrimaryCategory?: OrchestrationMusicGenre;
  musicGenrePrimarySubcategory?: OrchestrationMusicGenreSubcategory;
  musicGenreSecondaryCategory?: OrchestrationMusicGenre;
  musicGenreSecondarySubcategory?: OrchestrationMusicGenreSubcategory;
  play?: Play;
  createdAt: Date;
  updatedAt: Date;
  orchestrationBooks?: OrchestrationBook[];
};

export enum OrchestrationTypes {
  fullOrchestration = 'Full Orchestration',
  reducedOrchestration = 'Reduced Orchestration',
  pianoVocal = 'Piano/Vocal',
  rhythmSection = 'Rhythm Section',
  fullOrchestrationWithElectronicTracks = 'Full Orchestration with Electronic Tracks',
  reducedOrchestrationWithElectronicTracks = 'Reduced Orchestration with Electronic Tracks',
  pianoVocalWithElectronicTracks = 'Piano/Vocal with Electronic Tracks',
  rhythmSectionWithElectronicTracks = 'Rhythm Section with Electronic Tracks',
  electronicTracks = 'Electronic Tracks',
}

export enum OrchestrationMusicMoodStyle {
  acapella = 'Acapella',
  christian = 'Christian',
  featuringBass = 'Featuring Bass',
  featuringDrums = 'Featuring Drums',
  featuringGuitar = 'Featuring Guitar',
  featuringPiano = 'Featuring Piano',
  featuringSaxophone = 'Featuring saxophone',
  instrumental = 'Instrumental',
  moodAngry = 'Mood: Angry',
  moodBrooding = 'Mood: Brooding',
  moodChristmas = 'Mood: Christmas',
  moodDreamy = 'Mood: Dreamy',
  moodFun = 'Mood: Fun',
  moodFunny = 'Mood: Funny',
  moodIntellectual = 'Mood: Intellectual',
  moodPartyMusic = 'Mood: Party Music',
  moodPatriotic = 'Mood: Patriotic',
  moodQuirky = 'Mood: Quirky',
  moodSeasonal = 'Mood: Seasonal',
  moodSexMusic = 'Mood: Sex Music',
  moodUpbeat = 'Mood: Upbeat',
  moodVirtuoso = 'Mood: Virtuoso',
  moodWeird = 'Mood: Weird',
  outAndProud = 'Out-and-Proud',
  soloFemaleArtist = 'Solo Female Artist',
  soloMaleArtist = 'Solo Male Artist',
  soloNonBinaryArtist = 'Solo Non-Binary Artist',
  spiritual = 'Spiritual',
  typeAcoustic = 'Type: Acoustic',
  typeBackgroundMusic = 'Type: Background Music',
  typeCompilations = 'Type: Compilations',
  typeExperimental = 'Type: Experimental',
  typeImprovisational = 'Type: Improvisational',
  typeInstrumental = 'Type: Instrumental',
  typeLoFi = 'Type: Lo-Fi',
  typeLyrical = 'Type: Lyrical',
  typePolitical = 'Type: Political',
  typeSonic = 'Type: Sonic',
  typeSoundEffects = 'Type: Sound Effects',
  typeTributes = 'Type: Tributes',
  typeVocal = 'Type: Vocal',
}

export enum OrchestrationMusicStyle {
  avantGarde = 'Avant Garde',
  blues = 'Blues',
  brazilian = 'Brazilian',
  classical = 'Classical',
  country = 'Country',
  easyListening = 'Easy Listening',
  electronic = 'Electronic',
  folk = 'Folk',
  hipHopRap = 'Hip-Hop/Rap',
  holiday = 'Holiday',
  jazz = 'Jazz',
  kidsFamily = 'Kids/Family',
  latin = 'Latin',
  metalPunk = 'Metal/Punk',
  newAge = 'New Age',
  pop = 'Pop',
  rbSoul = 'R&B/Soul',
  reggae = 'Reggae',
  rock = 'Rock',
  spiritual = 'Spiritual',
  spokenWord = 'Spoken Word',
  world = 'World',
}

export enum OrchestrationMusicGenre {
  avantGarde = 'Avant Garde',
  blues = 'Blues',
  classical = 'Classical',
  country = 'Country',
  easyListening = 'Easy Listening',
  electronic = 'Electronic',
  folk = 'Folk',
  hipHopRap = 'Hip-Hop/Rap',
  jazz = 'Jazz',
  kidsFamily = 'Kids/Family',
  latin = 'Latin',
  metalPunk = 'Metal/Punk',
  moods = 'Moods',
  newAge = 'New Age',
  pop = 'Pop',
  reggae = 'Reggae',
  rock = 'Rock',
  spiritual = 'Spiritual',
  spokenWord = 'Spoken Word',
  urbanRB = 'Urban/R&B',
  world = 'World',
}

export type OrchestrationMusicGenreSubcategory =
  | AvantGardeGenre
  | BluesGenre
  | BrazilianGenre
  | ClassicalGenre
  | CountryGenre
  | EasyListeningGenre
  | ElectronicGenre
  | FolkGenre
  | HipHopRapGenre
  | JazzGenre
  | KidsFamilyGenre
  | LatinGenre
  | MetalPunk
  | NewAgeGenre
  | PopGenre
  | RBSoulGenre
  | ReggaeGenre
  | RockGenre
  | SpiritualGenre
  | SpokenWordGenre
  | WorldGenre;

export enum AvantGardeGenre {
  atonal = 'Atonal',
  avantAmericana = 'Avant-Americana',
  classicalAvantGarde = 'Classical Avant-Garde',
  computerMusic = 'Computer Music',
  electroAcoustic = 'Electro-Acoustic',
  electronicAvantGarde = 'Electronic Avant-Garde',
  experimental = 'Experimental',
  filmBaby = 'FilmBaby',
  freeImprovisation = 'Free Improvisation',
  lowerCaseMusic = 'Lower-case Music',
  microsound = 'Microsound',
  microtonal = 'Microtonal',
  mixedMedia = 'Mixed Media',
  modernComposition = 'Modern Composition',
  musiqueConcrete = 'Musique Concrete',
  noise = 'Noise',
  processGenerated = 'Process-Generated',
  psychedelia = 'Psychedelia',
  soundArt = 'Sound Art',
  soundCollage = 'Sound Collage',
  soundSculpture = 'Sound Sculpture',
  structuredImprovisation = 'Structured Improvisation',
  tapeMusic = 'Tape Music',
}

export enum BluesGenre {
  acousticBlues = 'Acoustic Blues',
  bluesGospel = 'Blues Gospel',
  bluesVocals = 'Blues Vocals',
  bluesRock = 'Blues-Rock',
  carolinaBeachShag = 'Carolina Beach / Shag',
  chicagoStyle = 'Chicago Style',
  classicFemaleBlues = 'Classic Female Blues',
  countryBlues = 'Country Blues',
  deltaStyle = 'Delta Style',
  detroitBlues = 'Detroit Blues',
  dirtyBlues = 'Dirty Blues',
  eastCoastBlues = 'East Coast Blues',
  electricBlues = 'Electric Blues',
  englishStyle = 'English Style',
  fingerPickedGuitar = 'Finger-Picked Guitar',
  folkBlues = 'Folk-Blues',
  funkyBlues = 'Funky Blues',
  guitarBlues = 'Guitar Blues',
  harmonicaBlues = 'Harmonica Blues',
  jazzyBlues = 'Jazzy Blues',
  jukeJointBlues = 'Juke Joint Blues',
  jumpBlues = 'Jump Blues',
  louisianaBlues = 'Louisiana Blues',
  mellowBlues = 'Mellow Blues',
  memphisBlues = 'Memphis Blues',
  newOrleansBlues = 'New Orleans Blues',
  newYorkBlues = 'New York Blues',
  pianoBlues = 'Piano Blues',
  piedmontBlues = 'Piedmont Blues',
  prewarBlues = 'Prewar Blues',
  rhythmBlues = 'Rhythm & Blues',
  rockinBlue = 'Rockin Blue',
  slideGuitarBlues = 'Slide Guitar Blues',
  soulBlues = 'Soul-Blues',
  stLouisBlues = 'St. Louis Blues',
  swampBlues = 'Swamp Blues',
  texasStyle = 'Texas Style',
  urbanBlues = 'Urban Blues',
  westCoastBlues = 'West Coast Blues',
  workSongs = 'Work Songs',
}

export enum BrazilianGenre {
  arrocha = 'Arrocha',
  axe = 'Axé',
  baiao = 'Baião',
  bandanejo = 'Bandanejo',
  batucada = 'Batucada',
  boiBumba = 'Boi-Bumba',
  bossaNova = 'Bossa Nova',
  brega = 'Brega',
  cacuria = 'Cacuriá',
  caipira = 'Caipira',
  cancioneiro = 'Cancioneiro',
  carimbo = 'Carimbó',
  chamame = 'Chamamé',
  chorinho = 'Chorinho',
  chula = 'Chula',
  ciranda = 'Ciranda',
  coco = 'Coco',
  embolada = 'Embolada',
  forro = 'Forró',
  frevo = 'Frevo',
  funkCarioca = 'Funk Carioca',
  gospel = 'Gospel',
  guitarrada = 'Guitarrada',
  maracatu = 'Maracatu',
  modaDeViola = 'Moda de Viola',
  mPB = 'MPB',
  pagode = 'Pagode',
  rasqueado = 'Rasqueado',
  samba = 'Samba',
  sambaDePartidoAlto = 'Samba de partido-alto',
  sambaCancao = 'Samba-canção',
  sambaEnredo = 'Samba-enredo',
  seresta = 'Seresta',
  sertanejo = 'Sertanejo',
  sertanejoUniversitário = 'Sertanejo Universitário',
  tcheMusic = 'Tchê Music',
  tecnoBrega = 'TecnoBrega',
  toada = 'Toada',
  trovaGalponeira = 'Trova Galponeira',
  vanerao = 'Vanerão',
}

export enum ClassicalGenre {
  arias = 'Arias',
  artSongs = 'Art Songs',
  bach = 'Bach',
  ballet = 'Ballet',
  bandMusic = 'Band Music',
  baroque = 'Baroque',
  brassQuintet = 'Brass quintet',
  cantata = 'Cantata',
  chamberMusic = 'Chamber Music',
  choralMusic = 'Choral Music',
  classicalEra = 'Classical era',
  concerto = 'Concerto',
  contemporary = 'Contemporary',
  earlyMusic = 'Early Music',
  filmMusic = 'Film Music',
  gregorianChant = 'Gregorian Chant',
  impressionism = 'Impressionism',
  keyboardMusic = 'Keyboard Music',
  madrigals = 'Madrigals',
  mass = 'Mass',
  medieval = 'Medieval',
  minimalism = 'Minimalism',
  modernist = 'Modernist',
  musicalTheater = 'Musical Theater',
  newAge = 'New Age',
  newMusicEnsemble = 'New Music Ensemble',
  opera = 'Opera',
  operetta = 'Operetta',
  oratorio = 'Oratorio',
  orchestral = 'Orchestral',
  organ = 'Organ',
  percussionEnsemble = 'Percussion Ensemble',
  postmodern = 'Postmodern',
  renaissance = 'Renaissance',
  rockOpera = 'Rock Opera',
  romanticEra = 'Romantic Era',
  serialist = 'Serialist',
  sinfonietta = 'Sinfonietta',
  tonePoem = 'Tone Poem',
  traditional = 'Traditional',
  twentiethCentury = 'Twentieth Century',
  vocalMusic = 'Vocal Music',
}

export enum CountryGenre {
  altCountry = 'Alt-Country',
  americana = 'Americana',
  bakersfieldSound = 'Bakersfield Sound',
  bluegrass = 'Bluegrass',
  contemporaryCountry = 'Contemporary Country',
  countryBlues = 'Country Blues',
  countryBoogie = 'Country Boogie',
  countryFolk = 'Country Folk',
  countryGospel = 'Country Gospel',
  countryPop = 'Country Pop',
  countryRock = 'Country Rock',
  countrypolitan = 'Countrypolitan',
  cowboy = 'Cowboy',
  honkyTonk = 'Honky Tonk',
  modernCountry = 'Modern Country',
  oldTimey = 'Old-Timey',
  outlawCountry = 'Outlaw Country',
  progressiveBluegrass = 'Progressive Bluegrass',
  progressiveCountry = 'Progressive Country',
  rodeo = 'Rodeo',
  squareDance = 'Square Dance',
  traditionalBluegrass = 'Traditional Bluegrass',
  traditionalCountry = 'Traditional Country',
  urbanCowboy = 'Urban Cowboy',
  western = 'Western',
  westernSwing = 'Western Swing',
  yodeling = 'Yodeling',
}

export enum EasyListeningGenre {
  adultContemporary = 'Adult Contemporary',
  americanPopularSong = 'American Popular Song',
  backgroundMusic = 'Background Music',
  ballads = 'Ballads',
  ballroom = 'Ballroom',
  barbershopQuartet = 'Barbershop Quartet',
  cabaret = 'Cabaret',
  classicalPop = 'Classical Pop',
  croonersVocals = 'Crooners/Vocals',
  easyPop = 'Easy Pop',
  exotica = 'Exotica',
  lounge = 'Lounge',
  mature = 'Mature',
  moodMusic = 'Mood Music',
  musicHall = 'Music Hall',
  musicalsBroadway = 'Musicals/Broadway',
  nostalgia = 'Nostalgia',
  orchestralPop = 'Orchestral Pop',
  polka = 'Polka',
  showTunes = 'Show Tunes',
  softRock = 'Soft Rock',
  spaceAgePop = 'Space Age Pop',
  tinPanAlley = 'Tin Pan Alley',
  torchSongs = 'Torch Songs',
  vaudeville = 'Vaudeville',
}
// ...
export enum FolkGenre {
  acidFolk = 'Acid Folk',
  alternativeFolk = 'Alternative Folk',
  angry = 'Angry',
  antiFolk = 'Anti-Folk',
  appalachianFolk = 'Appalachian Folk',
  britishFolk = 'British Folk',
  celticFolk = 'Celtic Folk',
  celticFusion = 'Celtic Fusion',
  celticPop = 'Celtic Pop',
  childrensFolk = "Children's Folk",
  contemporaryCeltic = 'Contemporary Celtic',
  contraDance = 'Contra Dance',
  drinkingSongs = 'Drinking Songs',
  fingerstyle = 'Fingerstyle',
  folkBlues = 'Folk Blues',
  folkPop = 'Folk Pop',
  folkJazz = 'Folk-Jazz',
  folkRock = 'Folk-Rock',
  freeFolk = 'Free-folk',
  gentle = 'Gentle',
  irishContemporary = 'Irish Contemporary',
  irishTraditional = 'Irish Traditional',
  jazzyFolk = 'Jazzy Folk',
  jugBand = 'Jug Band',
  likeAni = 'like Ani',
  likeJoni = 'like Joni',
  minstrel = 'Minstrel',
  modernFolk = 'Modern Folk',
  musette = 'Musette',
  pipeBand = 'Pipe Band',
  political = 'Political',
  politicalFolk = 'Political Folk',
  powerFolk = 'Power-folk',
  progressiveFolk = 'Progressive Folk',
  psychFolk = 'Psych-folk',
  scottishContemporary = 'Scottish Contemporary',
  scottishTraditional = 'Scottish Traditional',
  seaShanties = 'Sea Shanties',
  singerSongwriter = 'Singer/Songwriter',
  skiffle = 'Skiffle',
  stringBand = 'String Band',
  traditionalFolk = 'Traditional Folk',
  urbanFolk = 'Urban Folk',
}